requirements.txt
streamlit
pdf2image
python-docx
pytesseract
easyocr
opencv-python
pillow
pdfplumber
numpy
